test websockets on heroku

+++ src/main/resources/db/migration/V1__create_trades_table.sql
@@ -3,9 +3,9 @@ create table tradeMessages (
 	userId int not null,
 	currencyFrom varchar(3) not null,
 	currencyTo varchar(3) not null,
-	amountSell int not null,
-	amountBuy int not null,
-	rate float not null,
+	amountSell numeric(15,6) not null,
+	amountBuy numeric(15,6) not null,
+	rate numeric(15,6) not null,
 	timePlaced date not null,
 	originatingCountry varchar(2) not null
 );
